TASK 9 - FASTAPI,UI MODEL/backend/requirements.txt
fastapi
uvicorn
pandas
numpy
faiss-cpu
scikit-learn
python-multipart
